pyproject.toml
[project]
name = "bmds_ui"
license = {text = "MIT License"}
authors = [{name = "Andy Shapiro", email = "shapiro.andy@epa.gov"}]
readme = "README.md"
description = "US EPA Dose Response Modeling Software"
dynamic = ["version"]
keywords = [
  "dose response",
  "toxicology",
  "risk assessment",
]
classifiers = [
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
  # pybmds
  "pybmds~=24.1a0",

  # website
  "Django~=5.0.6",
  "django-anymail~=10.3",
  "django-redis~=5.4.0",
  "django-webpack-loader~=3.1.0",
  "djangorestframework~=3.15.1",
  "uritemplate~=4.1.1",
  "django-reversion~=5.0.12",
  "PyYAML~=6.0.1",

  # tasks
  "celery~=5.4.0",
  "billiard~=4.2.0",
  "kombu~=5.3.7",
  "plotly~=5.22.0",
  "redis~=5.0.4",

  # desktop
  "packaging~=24.0",
  "textual~=0.62.0",
  "whitenoise~=6.6.0",
]

[project.optional-dependencies]
dev = [
  # build
  "flit~=3.9.0",

  # debug
  "django-debug-toolbar~=4.3.0",
  "django_extensions==3.2.3",
  "django-browser-reload==1.12.1",

  # docs
  "Sphinx~=7.3.7",
  "myst-parser~=2.0.0",
  "pydata-sphinx-theme~=0.15.2",

  # tests
  "pytest~=8.2.1",
  "pytest-django~=4.8.0",
  "vcrpy~=6.0.1",
  "pytest-recording~=0.13.1",
  "pytest-asyncio==0.23.7",
  "coverage==7.5.1",

  # integration tests
  "playwright==1.44.0",
  "pytest-playwright==0.5.0",

  # lint and formatting tools
  "ruff~=0.4.4",
]
pg = [
  "psycopg2-binary~=2.9.9",
]
prod = [
  "gunicorn~=22.0.0",
  "sentry-sdk~=2.2.0",
]

[project.urls]
Source = "https://github.com/USEPA/bmds-online-private"

[project.scripts]
"manage.py" = "bmds_ui:manage"
manage = "bmds_ui:manage"
bmds = "bmds_ui.desktop.cli:main"
bmdsdesktop = "bmds_ui.desktop.cli:main"
"bmds_desktop" = "bmds_ui.desktop.cli:main"
"bmds-desktop" = "bmds_ui.desktop.cli:main"

[build-system]
requires = ["flit_core ~=3.9"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
include = ["bmds_ui/"]
exclude = ["bmds_ui/main/settings/local.py"]

[tool.coverage.run]
omit = [
    "../bmds/*",
    "../bmds-core/*",
    "./bmds_ui/main/settings/*",
    "./tests/*",
    "./venv/*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "S", "B", "T20", "RUF"]
ignore = ["E501", "B904", "B007", "S308", "S113", "S314"]
unfixable = ["F401", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["bmds_ui", "pybmds"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101", "S106"]
"scripts/*.py" = ["T201"]
"**/management/**.py" = ["T201"]
"**/migrations/**.py" = ["T201", "RUF012"]
"**/main/settings/**.py" = ["F403", "F405"]

[tool.pytest.ini_options]
addopts = "--ds=bmds_ui.main.settings.testing --reuse-db --nomigrations --record-mode=once"
testpaths = ["tests"]
python_files = ["test_*.py"]
